pyproject.toml
[project]
name = "digital_twin_like"
version = "0.1.0"
description = "digital-twin-like using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.186.1,<1.0.0",
    "openai-whisper>=20231117",
    "gTTS>=2.5.0",
    "pydub>=0.25.1",
    "sounddevice>=0.4.6",
    "soundfile>=0.12.1",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "python-dotenv>=1.0.0"
]

[project.scripts]
digital_twin_like = "digital_twin_like.main:run"
run_crew = "digital_twin_like.main:run"
train = "digital_twin_like.main:train"
replay = "digital_twin_like.main:replay"
test = "digital_twin_like.main:test"
voice_twin = "digital_twin_like.run_voice:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
